+++ CHANGELOG.md
@@ -4,12 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.1.3] - 2019-03-03
+### Added
+- Added Credential Manager and mimikatz/rubeus parser
+
 ### Changed
 - Updated Rubeus to latest commit
+- Changed Grunts to use CookieContainer WebClient for Cookie authentication
 
 ### Fixed
-- Re-added missing appsettings.json
+- Re-added missing appsettings.json, moved to Data folder
+- Check for initialized submodules
+- Fixed download task (missing folder)
 
 ## [v0.1.2] - 2019-02-14
 ### Added
@@ -42,3 +48,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [v0.1.1]: https://github.com/cobbr/Covenant/compare/v0.1...v0.1.1
 [v0.1.2]: https://github.com/cobbr/Covenant/compare/v0.1.1...v0.1.2
+[v0.1.3]: https://github.com/cobbr/Covenant/compare/v0.1.2...v0.1.3